# docs/swagger.yaml
basePath: /api
definitions:
  common.APIResponse:
    description: 所有 API 响应的统一包装格式，包含状态码、消息、数据和 timestamp
    properties:
      code:
        description: 状态码，200 表示成功，其他表示错误
        example: 200
        type: integer
      data:
        description: 响应数据，成功时返回业务数据，错误时为 null
      message:
        description: 消息提示，成功时为"success"，错误时为错误描述
        example: success
        type: string
      timestamp:
        description: Unix 时间戳（秒）
        example: 1717200000
        type: integer
    type: object
...
  controllers.SaveModuleConfigRequest:
    description: 保存模块配置的请求参数
    properties:
      content:
        description: 内容
        example: 主要内容
        type: string
      description:
        description: 描述
        example: 模块描述
        type: string
      enabled:
        description: 是否启用
        example: true
        type: boolean
      extraData:
        description: 额外数据（JSON 格式）
        example: '{"key":"value"}'
        type: string
      imagePath:
        description: 图片路径
        example: /uploads/images/banner.jpg
        type: string
      moduleName:
        description: 模块名称
        example: banner
        type: string
      sortOrder:
        description: 排序顺序
        example: 1
        type: integer
      subtitle:
        description: 副标题
        example: 副标题
        type: string
      title:
        description: 标题
        example: 标题
        type: string
    type: object
...
  controllers.UpdatePageConfigRequest:
    description: 更新页面配置的请求参数
    properties:
      configData:
        description: 配置数据（JSON 格式）
        example: '{"theme":"light"}'
        type: string
      pageName:
        description: 页面名称
        example: home
        type: string
    required:
    - pageName
    type: object
...
  gorm.DeletedAt:
    properties:
      time:
        type: string
      valid:
        description: Valid is true if Time is not NULL
        type: boolean
    type: object
...
  models.ModuleConfig:
    description: 存储各功能模块的配置信息，支持多语言
    properties:
      content:
        description: 内容（向后兼容）
        example: Content...
        type: string
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      description:
        description: 描述（向后兼容）
        example: Description
        type: string
      enContent:
        description: 英文内容
        example: Main content here...
        type: string
      enDescription:
        description: 英文描述
        example: Module description
        type: string
      enSubtitle:
        description: 英文副标题
        example: Professional medical device supplier
        type: string
      enTitle:
        description: 英文标题
        example: Welcome to our website
        type: string
      enabled:
        description: 是否启用
        example: true
        type: boolean
      extraData:
        description: 额外数据（JSON 格式）
        example: '{"key":"value"}'
        type: string
      id:
        type: integer
      imagePath:
        description: 图片路径
        example: /uploads/images/module.jpg
        type: string
      moduleName:
        description: 模块名称（唯一）
        example: banner
        type: string
      sortOrder:
        description: 排序顺序
        example: 1
        type: integer
      subtitle:
        description: 副标题（向后兼容）
        example: Subtitle
        type: string
      title:
        description: 标题（向后兼容）
        example: Title
        type: string
      updatedAt:
        type: string
      zhContent:
        description: 中文内容
        example: 这里是主要内容...
        type: string
      zhDescription:
        description: 中文描述
        example: 模块描述
        type: string
      zhSubtitle:
        description: 中文副标题
        example: 专业医疗器械供应商
        type: string
      zhTitle:
        description: 中文标题
        example: 欢迎来到我们的网站
        type: string
    type: object
  models.PageConfig:
    description: 存储各页面的配置数据
    properties:
      configData:
        description: 配置数据（JSON 格式）
        example: '{"theme":"light"}'
        type: string
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      id:
        type: integer
      pageName:
        description: 页面名称（唯一）
        example: home
        type: string
      updatedAt:
        type: string
    type: object
  models.SiteSetting:
    description: 存储站点配置项，键值对形式
    properties:
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      id:
        type: integer
      key:
        description: 设置键（唯一）
        example: site.name
        type: string
      updatedAt:
        type: string
      value:
        description: 设置值
        example: Shirley CMS
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@example.com
    name: API Support
  description: |-
    医疗器械内容管理系统 API 接口文档
    提供内容管理、模块配置、图片管理、多语言支持等功能
  title: Shirley CMS API
  version: "1.0"
paths:
  /api/admin/config:
    get:
      consumes:
      - application/json
      description: 获取指定页面或所有页面的配置信息
      parameters:
      - description: 页面名称，不传则返回所有页面配置
        in: query
        name: page
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 页面配置列表
          schema:
            allOf:
            - $ref: '#/definitions/common.APIResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.PageConfig'
                  type: array
              type: object
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/common.APIResponse'
      security:
      - APIKey: []
      summary: 获取页面配置
      tags:
      - 模块管理
    put:
      consumes:
      - application/json
      description: 更新指定页面的配置数据
      parameters:
      - description: 页面配置更新请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/controllers.UpdatePageConfigRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 更新成功
          schema:
            $ref: '#/definitions/common.APIResponse'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/common.APIResponse'
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/common.APIResponse'
      security:
      - APIKey: []
      summary: 更新页面配置
      tags:
      - 模块管理
...
  /api/admin/modules:
    get:
      consumes:
      - application/json
      description: 获取所有模块或指定模块的配置信息
      parameters:
      - description: 模块名称，不传则返回所有模块配置
        in: query
        name: module
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 模块配置列表
          schema:
            allOf:
            - $ref: '#/definitions/common.APIResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.ModuleConfig'
                  type: array
              type: object
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/common.APIResponse'
      security:
      - APIKey: []
      summary: 获取模块配置列表
      tags:
      - 模块管理
  /api/admin/modules/{name}:
    delete:
      consumes:
      - application/json
      description: 根据模块名称删除模块配置
      parameters:
      - description: 模块名称
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 删除成功
          schema:
            $ref: '#/definitions/common.APIResponse'
        "404":
          description: 模块配置不存在
          schema:
            $ref: '#/definitions/common.APIResponse'
      security:
      - APIKey: []
      summary: 删除模块配置
      tags:
      - 模块管理
    get:
      consumes:
      - application/json
      description: 根据模块名称获取模块配置详情
      parameters:
      - description: 模块名称
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 模块配置详情
          schema:
            allOf:
            - $ref: '#/definitions/common.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/models.ModuleConfig'
              type: object
        "404":
          description: 模块配置不存在
          schema:
            $ref: '#/definitions/common.APIResponse'
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/common.APIResponse'
      security:
      - APIKey: []
      summary: 获取单个模块配置
      tags:
      - 模块管理
    put:
      consumes:
      - application/json
      description: 创建或更新模块配置信息，支持表单和 JSON 格式
      parameters:
      - description: 模块名称
        in: path
        name: moduleName
        required: true
        type: string
      - description: 模块配置数据
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/controllers.SaveModuleConfigRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 保存成功
          schema:
            allOf:
            - $ref: '#/definitions/common.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/models.ModuleConfig'
              type: object
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/common.APIResponse'
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/common.APIResponse'
      security:
      - APIKey: []
      summary: 保存模块配置
      tags:
      - 模块管理
  /api/admin/modules/{name}/image:
    delete:
      consumes:
      - application/json
      description: 删除指定模块的关联图片
      parameters:
      - description: 模块名称
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 删除成功
          schema:
            $ref: '#/definitions/common.APIResponse'
        "404":
          description: 模块图片不存在
          schema:
            $ref: '#/definitions/common.APIResponse'
      security:
      - APIKey: []
      summary: 删除模块图片
      tags:
      - 模块管理
  /api/admin/site-settings:
    get:
      consumes:
      - application/json
      description: 获取所有站点配置项
      produces:
      - application/json
      responses:
        "200":
          description: 站点设置列表
          schema:
            allOf:
            - $ref: '#/definitions/common.APIResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.SiteSetting'
                  type: array
              type: object
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/common.APIResponse'
      security:
      - APIKey: []
      summary: 获取站点设置
      tags:
      - 站点设置
    post:
      consumes:
      - multipart/form-data
      description: 保存站点标题等配置项
      parameters:
      - description: 中文站点标题
        in: formData
        name: zhSiteTitle
        type: string
      - description: 英文站点标题
        in: formData
        name: enSiteTitle
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 保存成功
          schema:
            allOf:
            - $ref: '#/definitions/common.APIResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.SiteSetting'
                  type: array
              type: object
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/common.APIResponse'
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/common.APIResponse'
      security:
      - APIKey: []
      summary: 保存站点设置
      tags:
      - 站点设置
...
  /api/public/config:
    get:
      consumes:
      - application/json
      description: 获取所有页面的公开配置信息（无需认证）
      produces:
      - application/json
      responses:
        "200":
          description: 页面配置列表
          schema:
            allOf:
            - $ref: '#/definitions/common.APIResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.PageConfig'
                  type: array
              type: object
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/common.APIResponse'
      summary: 获取公开页面配置
      tags:
      - 公开接口
...
  /api/public/modules:
    get:
      consumes:
      - application/json
      description: 获取所有模块的公开配置信息（无需认证）
      produces:
      - application/json
      responses:
        "200":
          description: 模块配置列表
          schema:
            allOf:
            - $ref: '#/definitions/common.APIResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.ModuleConfig'
                  type: array
              type: object
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/common.APIResponse'
      summary: 获取公开模块配置
      tags:
      - 公开接口
  /api/public/site-settings:
    get:
      consumes:
      - application/json
      description: 获取站点配置的公开信息（无需认证）
      produces:
      - application/json
      responses:
        "200":
          description: 站点设置列表
          schema:
            allOf:
            - $ref: '#/definitions/common.APIResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.SiteSetting'
                  type: array
              type: object
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/common.APIResponse'
      summary: 获取公开站点设置
      tags:
      - 公开接口
securityDefinitions:
  APIKey:
    description: JWT Token 认证，格式：Bearer {token}
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"